Move client components
