+++ src/app.tsx
@@ -7,9 +7,9 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { ResourcePage } from './components/ResourcePage';
 import { NavBar } from './components/NavBar';
 
-import { ClientsPage } from './components/client/ClientsPage';
-import { ClientNewPage } from './components/client/ClientNewPage';
-import { ClientPage } from './components/client/ClientPage';
+import { ClientsPage } from './pages/clients/ClientsPage';
+import { ClientNewPage } from './pages/clients/new/ClientNewPage';
+import { ClientPage } from './pages/clients/client/ClientPage';
 
 import { PersonPage } from './components/person/PersonPage';
 import { PersonNewPage } from './components/person/PersonNewPage';
